Ensure the 'Counterparty.Website' field has a protocol attached (#327)
Add parsing/checks to ensure a protocol schema like 'https://' is on every counterparty website when creating/updating the counterparties or importing them from a Daybreak list. Added a function to the model.Counterparty called NormalizedWebsite() which returns a string that will be parsed and have a schema (or an error).